src/tg_bot/requirements.txt
piexif
pillow
requests
